Move end_screen class and functions to their dedicated files
Add "actually_dead" debug argument for end screen handling
